Move calculateLabels into separate file
